# api.yaml
openapi: 3.0.0
info:
  title: MediCore API
  description: Healthcare management system API for patients, doctors, and administrators
  version: 1.0.0
  contact:
    name: MediCore Support
servers:
  - url: http://localhost:5000/api
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CaseNoteRequest:
      type: object
      required:
        - condition
        - treatment
      properties:
        condition:
          type: string
        treatment:
          type: string
        remarks:
          type: string
        booking_id:
          type: integer
security:
  - bearerAuth: []
paths:
...
  /doctor/profile:
    get:
      tags:
        - Doctor
      summary: Get doctor profile
      responses:
        '200':
          description: Doctor profile
        '404':
          description: Profile not found
    put:
      tags:
        - Doctor
      summary: Update doctor profile
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                phone:
                  type: string
                expertise:
                  type: string
                years_exp:
                  type: integer
                consult_fee:
                  type: number
                schedule_text:
                  type: string
                about:
                  type: string
      responses:
        '200':
          description: Profile updated
...
  /doctor/bookings/{id}/approve:
    put:
      tags:
        - Doctor
      summary: Approve booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Booking approved
        '403':
          description: Unauthorized
  /doctor/bookings/{id}/complete:
    put:
      tags:
        - Doctor
      summary: Mark booking as completed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Booking completed
  /doctor/bookings/{id}/cancel:
    put:
      tags:
        - Doctor
      summary: Cancel booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Booking cancelled
...
  /doctor/notes/{cid}:
    post:
      tags:
        - Doctor
      summary: Add case note for client
      parameters:
        - name: cid
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CaseNoteRequest'
      responses:
        '201':
          description: Note added
        '400':
          description: Missing required fields
  /doctor/schedule:
    get:
      tags:
        - Doctor
      summary: Get doctor schedule (7 days)
      responses:
        '200':
          description: Schedule for next 7 days
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    date:
                      type: string
                      format: date
                    is_open:
                      type: boolean
    post:
      tags:
        - Doctor
      summary: Update doctor schedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                properties:
                  date:
                    type: string
                    format: date
                  is_open:
                    type: boolean
      responses:
        '200':
          description: Schedule updated
...
  /admin/reports/monthly:
    get:
      tags:
        - Admin
      summary: Get monthly report
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Monthly report data
  /admin/physicians:
    get:
      tags:
        - Admin
      summary: Get all physicians
      parameters:
        - name: q
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of physicians
    post:
      tags:
        - Admin
      summary: Add new physician
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - name
                - email
                - password
                - expertise
              properties:
                username:
                  type: string
                name:
                  type: string
                email:
                  type: string
                password:
                  type: string
                phone:
                  type: string
                expertise:
                  type: string
                years_exp:
                  type: integer
                consult_fee:
                  type: number
                schedule_text:
                  type: string
                division_id:
                  type: integer
                about:
                  type: string
      responses:
        '201':
          description: Physician added
        '400':
          description: Validation error
  /admin/physicians/{id}:
    put:
      tags:
        - Admin
      summary: Update physician
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Physician updated
    delete:
      tags:
        - Admin
      summary: Delete physician
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Physician deleted
  /admin/clients:
    get:
      tags:
        - Admin
      summary: Get all clients
      parameters:
        - name: q
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of clients
    post:
      tags:
        - Admin
      summary: Add new client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - name
                - email
                - password
      responses:
        '201':
          description: Client added
  /admin/clients/{id}:
    get:
      tags:
        - Admin
      summary: View client details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Client details with bookings and notes
    put:
      tags:
        - Admin
      summary: Update client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Client updated
    delete:
      tags:
        - Admin
      summary: Delete client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Client deleted
...